Update README.md
My experience with GitHub so far has been quite seamless. Cloud computing is an amazing technology that allows you to save data automatically onto a server wherever you are located. This allows for access to work without having to be on that exact local machine. Making it super simple to switch between devices when im on the go. 

The interaction with the client and the server is seamless too. Every segment of the website loads with no issues and the connection between them has been outstanding. I know that my work will be saved without having to put any extra stress into backing data up on an external.